+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # **Upcoming release**
 
+- Remove site-packages from packages search tree (@tkrabel)
+
+# Release 1.11.0
+
 - #710, #561 Implement `except*` syntax (@lieryan)
-- #711 Allow building documentation without having rope module installed (@kloczek)
-- #722 Speed up module cache generation by factor of 2 (@tkrabel)
+- #711 allow building documentation without having rope module installed (@kloczek)
+- #719 Allows the in-memory db to be shared across threads (@tkrabel)
+- #720 create one sqlite3.Connection per thread using a thread local (@tkrabel)
+- #715 change AutoImport's `get_modules` to be case sensitive (@bagel897)
 
 # Release 1.10.0